feat(build): give the source copy gate the dist gate's normaliser (wayfinder #68)
#68 asked whether checks.yml should build, so the dist copy gate could run on
pull requests. Answered no: the gap that argued for it closes more cheaply, and
more completely, on the source side.

Measured first. The dist gate's unique reach over the source gate turned out to
be one surface — prose typed straight into .astro markup — via three ordinary
authoring accidents the line matcher could not see:

  <strong>no</strong> surprises      split across a tag
  at&nbsp;cost                       HTML entity
  no / surprises                     wrapped mid-phrase across a newline

All three failed only after a merge, in the dist gate, on main. Everything else
the dist gate saw, the source gate already saw: every authored string lives in
src/i18n/*.ts, which it scans, so #57's "the built HTML would have caught almost
nothing" was an argument about which dist files to read, not evidence that dist
sees strings source cannot.

So the normaliser moves to scripts/copy-gate-normalise.mjs and both gates import
it — one pattern list AND one normaliser, for the reason #57 gave for the list: a
transform present in one gate and absent from the other reads as covered when it
is not. The source gate now reads each file twice, raw per line and whole-file
normalised, and a normalised hit still reports a line (a range when the phrase
crosses one) plus the view that found it, because the position map makes a match
a span of original bytes rather than an offset into a derived string.

Two things decided rather than defaulted:

  - The source gate DECLINES the two views that can fabricate a phrase
    (tag-to-a-space, unknown-reference-to-a-space). Its allowlist is a per-line
    pragma a human writes, and a pragma blessing a claim nobody made teaches the
    next reader that the gate cries wolf. The case it gives up — a phrase whose
    only separator is a tag boundary — stays the dist gate's, which blesses
    phantoms in a list rather than in the copy.
  - A pragma covers the line a match STARTS on, not any line in its span. The
    alternative is the laundering shape an earlier review already closed here.

The dist gate keeps its job and its header now says what is actually left to it:
vendored files this repo does not author, text the build introduces, literal
concatenation Rollup folds ("insur" + "ance"), the fabricating views, and being
the only layer that measures the promise itself rather than a proxy.

checks.yml stays dependency-free — ~50ms for this gate, ~12s for the job. It
records the question and the answer, so nobody re-asks it from scratch.

Verified: 32 controls, 0 failures — every split, entity, escape, invisible
character and wrapped phrase fails; accented Spanish, the `passthrough`
identifier, "No surge today", a double-encoded reference, a TS regex's \b
escapes, `a < b`, and a blanked comment between two lines do not fabricate; the
pragma pair proves start-line coverage and end-line non-coverage. The dist gate
was proved behaviour-preserving by diffing old against new over a synthetic
dist/ exercising every view, allowlist and count path — identical but for one
renamed view label (as-shipped -> plain, since a source file is not shipped).
The real tree stays green with no new copy-gate allowances; one extra `allowed:`
line appears, "seguro" on legalCopy.ts:337, the second word of §5's
(seguros?|aseguranza|póliza) alternation that pass 1's single line.match never
reported. Four places calling this gate "a line matcher" were made true again.

astro build is unverified for want of a registry token, as in #59 — but nothing
here needs one, which is the point.
