+++ CLAUDE.md
@@ -18,10 +18,10 @@ npm run preview    # Preview production build locally
 No test runner and no linter are configured. `npm run build` is the only automated verification gate. It runs five things in order, and any one of them fails the build:
 
 1. **`scripts/check-route-parity.mjs`** — every route under `src/pages` has its `/es/` twin. `404` and `redirect` are exempt by name (single file, language switched client-side). Routes whose twin is not built yet sit in a `PENDING` map naming the ticket that retires them; an entry whose twin now exists fails, so the list cannot go stale.
-2. **`scripts/check-copy-gate.mjs`** — gated and never-claimed strings (`docs/copy-map.md` §5, on the unmerged `copy-map/en-es` branch) must not reach the site. Scans `src/` and `public/` and fails on a Greek or Cyrillic letter, which in EN/ES copy is a homoglyph hiding a gated word. It is a line matcher over **source** text, so a phrase broken across a tag, an entity or a newline slips it — that is item 5's job. Its value is the file-and-line failure at the point of authorship, and it owns the per-line `copy-gate-allow` escape hatch; see the script header.
+2. **`scripts/check-copy-gate.mjs`** — gated and never-claimed strings (`docs/copy-map.md` §5, on the unmerged `copy-map/en-es` branch) must not reach the site. Scans `src/` and `public/` and fails on a Greek or Cyrillic letter, which in EN/ES copy is a homoglyph hiding a gated word. It reads each file **twice**: once raw, line by line, and once normalised through `scripts/copy-gate-normalise.mjs` — the same reading machine item 5 uses (#68), so a phrase broken across a tag, an entity, an escape or a newline fails here, at the point of authorship, rather than after a merge. A normalised hit still reports a line, a range when the phrase crosses one, and names the view that found it. It declines the two views that can *invent* a phrase, so no `copy-gate-allow` pragma ever has to bless a phantom; those stay item 5's. Its value is the file-and-line failure at authorship time, and it owns the per-line `copy-gate-allow` escape hatch — which covers the line a match **starts** on; see the script header.
 3. **`scripts/check-env.mjs`** — the six `PUBLIC_*` variables are present, non-empty and printable ASCII, and `PUBLIC_API_URL` ends with its slash. The printable-ASCII rule is a whitelist on purpose: an invisible character (a paste artifact like U+200E or U+00AD) survives `dotenv` and reaches the bundle, where it stops a URL parsing at all, and a blocklist of such characters cannot be finished. Astro inlines these at build time, so one that is not there becomes a falsy literal — `undefined` where the key is absent, `""` where it is present but empty — and Rollup then folds away the branch that tested it: an empty `PUBLIC_API_URL` ships a pre-registration form whose submit path no longer exists in the bundle, off a green build. The production trigger is the *empty* case, not the undefined one — a renamed or deleted GitHub Secret interpolates to `""` in `deploy-all.yml`'s "Create env file" step. It reads the environment through Vite's own `loadEnv`, so it sees exactly what the build will see (`.env` → `.env.local` → `.env.production` → `.env.production.local` → `process.env`, with an empty value in a later source blanking a non-empty one from an earlier). It needs both `npm ci` and the secrets, so it cannot join `checks.yml`. Run it alone with `npm run check:env`.
 4. **`astro check`** — a type error fails the build.
-5. **`scripts/check-dist-copy-gate.mjs`** — the same §5 list (both gates import `scripts/copy-gate-patterns.mjs`; one list, never two) run over `dist/` **after** `astro build`, with character references decoded, invisible characters removed, JS escapes resolved, whitespace collapsed and three views of each HTML file matched. This is the layer that measures the actual promise — a forbidden claim must not reach the shipped site — and it is not optional for `/fees`, whose copy is rendered client-side and so exists only in the JS bundle, never in the built HTML. Source pragmas are stripped by the build and cannot be seen from here, so it carries its **own** allowlist keyed to the hash-free path, the matched text and an exact count; an entry that stops matching, or matches a different number of times, fails. Run it alone with `npm run check:dist` (needs a `dist/`).
+5. **`scripts/check-dist-copy-gate.mjs`** — the same §5 list run over `dist/` **after** `astro build`, through six views of each markup file (two of which the source gate declines because they can fabricate). Both gates import **one** pattern list, `scripts/copy-gate-patterns.mjs`, and **one** normaliser, `scripts/copy-gate-normalise.mjs` — never two of either, because a pattern or a transform present in one gate and absent from the other reads as covered when it is not. This is the layer that measures the actual promise — a forbidden claim must not reach the shipped site. Since #68 gave the source gate the same reading, what only this gate can still see is narrower and worth naming: **vendored** files this repo does not author, text the **build** introduces, and literal concatenation Rollup folds (`"insur" + "ance"`). Source pragmas are stripped by the build and cannot be seen from here, so it carries its **own** allowlist keyed to the hash-free path, the matched text and an exact count; an entry that stops matching, or matches a different number of times, fails. Run it alone with `npm run check:dist` (needs a `dist/`).
 
 `.github/workflows/checks.yml` runs items 1–2 on every pull request — it is dependency-free by design, and items 3–5 all need `npm ci` against the private registry (item 3 reads the environment through Astro's own Vite), with item 3 additionally needing the secrets a PR does not have. `deploy-all.yml` runs the full chain on `main` and gates deployment on it, so items 3 and 5 are pre-deploy backstops rather than PR-time ones.
