+++ scripts/check-dist-copy-gate.mjs
@@ -1,11 +1,24 @@
 // Copy gate, second layer — the same §5 pattern list, run over dist/ after the
 // build. Wayfinder #57.
 //
-// WHY A SECOND GATE. check-copy-gate.mjs matches lines of source text, so it is
-// blind to everything the build does to that text: a phrase broken across a tag,
-// an HTML entity or a newline slips it, and ordinary Prettier wrapping produces
-// that by accident. This one measures the promise itself — a forbidden claim must
-// not reach the shipped site — by reading what is actually shipped.
+// WHY A SECOND GATE. This one measures the promise itself — a forbidden claim
+// must not reach the shipped site — by reading what is actually shipped.
+//
+// That is a smaller gap than it was. #57 wrote this gate because the source gate
+// was a line matcher over source text, blind to a phrase broken across a tag, an
+// HTML entity or a newline; #68 then moved this file's normaliser into
+// scripts/copy-gate-normalise.mjs and gave the source gate the same reading, so
+// all three of those now fail at the point of authorship instead of after a
+// merge. What is left here, and it is not nothing:
+//   - files this repo does not author — vendored bundles and brand assets, which
+//     are in dist/ and not in src/.
+//   - text the BUILD introduces, rather than the author.
+//   - literal concatenation Rollup folds: "insur" + "ance" is two harmless
+//     strings in the source and one word in the bundle.
+//   - the two views that can fabricate, which the source gate declines so that no
+//     per-line pragma ever has to bless a phantom.
+// And it is the only layer that measures the thing itself rather than a proxy for
+// it, which is why it gates the deploy.
 //
 // It does not replace the first gate. The source gate fails at the point of
 // authorship with a file and a line number, runs on every PR without a build or
@@ -41,43 +54,13 @@
 // single "×1" allowance passed two shipped claims on the one page this gate
 // exists to protect. Spans cannot be fooled that way.
 //
-// The views, and what each is for:
-//   - tags kept          the plain reading; also the only view that sees
-//                        attribute text (alt, title, content), since stripping a
-//                        tag takes its attributes with it.
-//   - tags removed       catches "Insur<span>ance</span>". It cannot fabricate a
-//                        multi-word phrase: removing a tag removes the gap, so
-//                        "<li>zero</li><li>markup</li>" becomes "zeromarkup".
-//   - tags to a space    catches a phrase split across two elements. This is the
-//                        one view that CAN fabricate — "<td>no</td>
-//                        <td>surprises</td>" reads as two cells — so every hit
-//                        names the view that produced it, and a fabricated one is
-//                        blessed once in ALLOWED below.
-//   - unknown references blanked / spaced
-//                        the NAMED table below cannot hold all ~2200 HTML5
-//                        entities. Rather than pretend, anything it does not know
-//                        is ALSO tried as "" and as " ", so "no&NewLine;surprises"
-//                        is caught without needing &NewLine; in the table.
-//   - escapes decoded    \uXXXX, \xXX, \u{...} and CSS's "\69 " form. Applied to
-//                        markup as well as scripts, because an inline <script> in
-//                        a built page is markup by extension and a bundle by
-//                        content.
-//
-// Applied in every view, because each is exactly what a browser does:
-//   - character references decoded in ONE left-to-right pass. One pass is the
-//     point: "&amp;#36;" renders as the literal text "&#36;", not "$", and a
-//     second pass would invent a hit the reader never sees.
-//   - invisible characters (zero-width space/joiner, soft hyphen, BOM) removed —
-//     they exist only to split a word without showing it.
-//   - whitespace collapsed to single spaces, so a newline inside a phrase — the
-//     accidental case — reads as the space it renders as.
-//
-// NOT applied: homoglyph folding. Folding needs a confusables table, and a wrong
-// entry in it invents failures in Spanish copy, which is full of legitimate
-// non-ASCII. A confusable is instead caught where it is introduced —
-// check-copy-gate.mjs fails on Cyrillic, Latin-lookalike Greek, fullwidth and
-// mathematical letters in source, which is unambiguous there because no vendored
-// code is in scope.
+// The reading machine itself — the views, the character-reference decoding, the
+// invisible-character removal, the whitespace collapse and the position map that
+// makes spans possible — lives in scripts/copy-gate-normalise.mjs, shared with
+// the source gate (#68). One list AND one normaliser: a transform present in one
+// gate and absent from the other reads as covered when it is not. Read that file
+// for what each view is for and why this gate asks for the two that can
+// fabricate while the source gate does not.
 //
 // ---------------------------------------------------------------------------
 // WHAT IT STILL CANNOT DO:
@@ -104,6 +87,7 @@ import { readdirSync, readFileSync, statSync } from "node:fs";
 import { basename, join } from "node:path";
 
 import { PATTERNS } from "./copy-gate-patterns.mjs";
+import { countOccurrences, matchesIn, viewsOf } from "./copy-gate-normalise.mjs";
 
 const ROOT = "dist";
 // Case-insensitive: a page served as .HTML is still a page. GitHub Pages serves
@@ -185,162 +169,6 @@ const ALLOWED = [
 const ISSUE_OR_VENDOR = /#\d+(?!\w)|^vendored:/;
 
 // ---------------------------------------------------------------------------
-// Normalisation
-
-const REFERENCE = /^&(#[xX][0-9a-fA-F]{1,6}|#\d{1,7}|[a-zA-Z][a-zA-Z0-9]{1,31});/;
-
-// The named references Astro emits, plus the ones that would be used to split or
-// disguise a word. This is NOT the full HTML5 set and does not try to be — the
-// "unknown reference blanked / spaced" views cover whatever is missing.
-const NAMED = {
-  amp: "&", lt: "<", gt: ">", quot: '"', apos: "'",
-  nbsp: " ", ensp: " ", emsp: " ", thinsp: " ", hairsp: " ", puncsp: " ",
-  newline: "\n", tab: "\t",
-  shy: "", zwnj: "", zwj: "", lrm: "", rlm: "",
-  mdash: "—", ndash: "–", hellip: "…", middot: "·",
-  lsquo: "‘", rsquo: "’", ldquo: "“", rdquo: "”",
-  dollar: "$", cent: "¢", pound: "£", euro: "€", yen: "¥",
-  copy: "©", reg: "®", trade: "™", deg: "°", sect: "§", para: "¶",
-};
-
-// Written as escapes, not as the characters themselves: these are invisible, so a
-// literal class here would be unreviewable, would make this file read as binary
-// to grep and file(1), and one lost byte in an editor would silently narrow the
-// gate. Soft hyphen, the zero-width/bidi block, word joiner, BOM.
-const INVISIBLE = new Set([
-  "\u00AD", "\u200B", "\u200C", "\u200D", "\u200E", "\u200F", "\u2060", "\uFEFF",
-]);
-
-const cp = (n) => (Number.isFinite(n) && n >= 0 && n <= 0x10ffff ? String.fromCodePoint(n) : "");
-
-/** The decoded value of one character reference, or null if this table cannot
- *  resolve it. Numeric references always resolve; named ones may not. */
-function decodeReference(body) {
-  if (body[0] === "#") {
-    const hex = body[1] === "x" || body[1] === "X";
-    const n = parseInt(hex ? body.slice(2) : body.slice(1), hex ? 16 : 10);
-    return Number.isNaN(n) ? null : cp(n);
-  }
-  const named = NAMED[body.toLowerCase()];
-  return named === undefined ? null : named;
-}
-
-/** A JS or CSS escape at `raw[i]`, as [decodedText, consumedLength], or null.
- *  CSS spells the same idea "\69 " — 1-6 hex digits and an optional trailing
- *  space that is part of the escape rather than of the text. */
-function decodeEscape(raw, i) {
-  const braced = /^\\u\{([0-9a-fA-F]{1,6})\}/.exec(raw.slice(i, i + 10));
-  if (braced) return [cp(parseInt(braced[1], 16)), braced[0].length];
-
-  const u = /^\\u([0-9a-fA-F]{4})/.exec(raw.slice(i, i + 6));
-  if (u) return [cp(parseInt(u[1], 16)), u[0].length];
-
-  const x = /^\\x([0-9a-fA-F]{2})/.exec(raw.slice(i, i + 4));
-  if (x) return [cp(parseInt(x[1], 16)), x[0].length];
-
-  const css = /^\\([0-9a-fA-F]{1,6})[ \t\n]?/.exec(raw.slice(i, i + 9));
-  if (css) return [cp(parseInt(css[1], 16)), css[0].length];
-
-  return null;
-}
-
-/**
- * One reading of `raw`.
- *
- * Returns { text, map } where map[n] is the offset in `raw` that text[n] came
- * from, so a match in `text` can be reported as a span of original bytes. Every
- * view of a file maps back to the same coordinates, which is what lets
- * overlapping matches from different views be recognised as one occurrence.
- */
-function view(raw, { tags = "keep", unknownReference = "keep", escapes = false }) {
-  const out = [];
-  const map = [];
-  let pendingSpace = -1;
-
-  const emit = (ch, at) => {
-    if (INVISIBLE.has(ch)) return;
-    if (/\s/.test(ch)) {
-      if (pendingSpace < 0) pendingSpace = at;
-      return;
-    }
-    if (pendingSpace >= 0) {
-      out.push(" ");
-      map.push(pendingSpace);
-      pendingSpace = -1;
-    }
-    out.push(ch);
-    map.push(at);
-  };
-  const emitAll = (s, at) => {
-    for (const ch of s) emit(ch, at);
-  };
-
-  for (let i = 0; i < raw.length; ) {
-    const ch = raw[i];
-
-    if (tags !== "keep" && ch === "<") {
-      const end = raw.indexOf(">", i);
-      if (end !== -1) {
-        if (tags === "space") emit(" ", i);
-        i = end + 1;
-        continue;
-      }
-    }
-
-    if (ch === "&") {
-      const hit = REFERENCE.exec(raw.slice(i, i + 36));
-      if (hit) {
-        const decoded = decodeReference(hit[1]);
-        if (decoded !== null) emitAll(decoded, i);
-        else if (unknownReference === "blank") void 0;
-        else if (unknownReference === "space") emit(" ", i);
-        else emitAll(hit[0], i);
-        i += hit[0].length;
-        continue;
-      }
-    }
-
-    if (escapes && ch === "\\") {
-      const hit = decodeEscape(raw, i);
-      if (hit) {
-        emitAll(hit[0], i);
-        i += hit[1];
-        continue;
-      }
-    }
-
-    emit(ch, i);
-    i++;
-  }
-
-  return { text: out.join(""), map };
-}
-
-/** Every reading of one file. Markup gets the tag treatments; everything else is
- *  read once plain and once with escapes resolved. Character references are
- *  decoded everywhere, because a bundle that writes them into innerHTML
- *  publishes them. */
-function viewsOf(path, raw) {
-  const named = [];
-  const add = (name, opts) => named.push({ name, ...view(raw, opts) });
-
-  if (MARKUP_EXT.test(path)) {
-    add("as-shipped", { tags: "keep" });
-    add("tags-removed", { tags: "drop" });
-    add("tags-as-space", { tags: "space" });
-    add("escapes-decoded", { tags: "keep", escapes: true });
-    add("entities-blanked", { tags: "drop", unknownReference: "blank" });
-    add("entities-spaced", { tags: "space", unknownReference: "space" });
-    return named;
-  }
-
-  add("as-shipped", {});
-  add("escapes-decoded", { escapes: true });
-  add("entities-spaced", { unknownReference: "space" });
-  return named;
-}
-
-// ---------------------------------------------------------------------------
 
 function walk(dir) {
   return readdirSync(dir, { withFileTypes: true }).flatMap((entry) => {
@@ -368,23 +196,6 @@ const extensionOf = (path) => {
   if (dot <= 0) return "(no extension)";
   return name.slice(dot);
 };
-
-/** Merge overlapping spans. Two matches that overlap in the original bytes are
- *  the same claim seen through two views; disjoint ones are separate claims. */
-function countOccurrences(spans) {
-  const sorted = [...spans].sort((a, b) => a[0] - b[0] || a[1] - b[1]);
-  let count = 0;
-  let end = -1;
-  for (const [start, stop] of sorted) {
-    if (start >= end) {
-      count++;
-      end = stop;
-    } else if (stop > end) {
-      end = stop;
-    }
-  }
-  return count;
-}
 
 let everything;
 try {
@@ -430,28 +241,28 @@ for (const path of files) {
 
 for (const path of files) {
   const key = keyOf(path);
-  const views = viewsOf(path, readFileSync(path, "utf8"));
+  // includeFabricating: this gate takes the two views that can invent a phrase —
+  // it measures the shipped promise and would rather fail on a phantom than miss
+  // a real claim, and ALLOWED is where each phantom gets blessed. The source gate
+  // passes false; copy-gate-normalise.mjs explains the asymmetry.
+  const views = viewsOf(readFileSync(path, "utf8"), {
+    markup: MARKUP_EXT.test(path),
+    includeFabricating: true,
+  });
 
   // (pattern, matched text) -> { spans, view }. Keyed by the text as well as the
   // pattern, because one alternation matches several words — §5's
   // /\b(seguros?|aseguranza|p[óo]liza)\b/ finds both halves of "póliza de
   // seguro" — and collapsing those into one entry would make the allowlist
   // unreadable. Spans are original-byte ranges, so within a key a match found by
-  // several views is one occurrence, not several.
+  // several views is one occurrence, not several; `view` names the first one
+  // that saw it, which is why matchesIn walks views outermost.
   const found = new Map();
-  for (const { name, text, map } of views) {
-    for (const [at, { re }] of PATTERNS.entries()) {
-      const global = new RegExp(re.source, re.flags.includes("g") ? re.flags : `${re.flags}g`);
-      for (const m of text.matchAll(global)) {
-        if (!m[0].length) continue;
-        const matched = m[0].toLowerCase();
-        const id = `${at}\u0000${matched}`;
-        const span = [map[m.index], map[m.index + m[0].length - 1] + 1];
-        const seen = found.get(id) ?? { at, spans: [], text: matched, view: name };
-        seen.spans.push(span);
-        found.set(id, seen);
-      }
-    }
+  for (const { at, text, view: name, span } of matchesIn(views, PATTERNS)) {
+    const id = `${at}\u0000${text}`;
+    const seen = found.get(id) ?? { at, spans: [], text, view: name };
+    seen.spans.push(span);
+    found.set(id, seen);
   }
 
   for (const { at, spans, text, view: viewName } of found.values()) {
